refactor: use schemaless URL to fetch SDK
This is not a best practice, but I'd rather not hard-code an HTTPS URL
here because I know my localhost doesn't have a valid certificate.

This will stop this error in the Chrome console:

    The page at 'https://wincent.github.io/remote-app-test/' was loaded
    over HTTPS, but requested an insecure script
    'http://0.0.0.0:8080/o/js/resolved-module/remote-web-app-client-js@1.0.0/js/index.js'.
    This request has been blocked; the content must be served over HTTPS.

But it will yield a new error about no valid certificate, which I am
hoping I can suppress via devtool tweaking.